+++ shared/types/src/elastic/contributions.ts
@@ -1,9 +1,4 @@
-import {
-  ContributionContentType,
-  ContributionHighlight,
-  ContributionLinkedContent,
-  ContributionRef,
-} from "../hasura";
+import { ContributionDocumentJson, ContributionHighlight } from "../hasura";
 import { Breadcrumb, DocumentElasticWithSource } from "./common";
 
 export interface ContributionConventionnelInfos {
@@ -21,44 +16,24 @@ export interface ContributionMetadata {
   text: string;
 }
 
-type ContributionDocumentJsonBasic = {
-  references: ContributionRef[];
-  contentType: ContributionContentType;
-  linkedContent: ContributionLinkedContent[];
-  questionIndex: number;
-  questionName: string;
-  questionId: string;
-  description: string;
-  idcc: string; // 0000 pour la générique, impossible d'être nulle
-};
-
-type ContributionDocumentJsonContent = ContributionDocumentJsonBasic & {
-  type: "content";
+export interface ContributionContentBase {
   content: string;
-};
+}
 
-type ContributionDocumentJsonGenericNoCDT = ContributionDocumentJsonBasic & {
-  type: "generic-no-cdt";
+export interface ContributionFicheSpContent {
+  url: string;
+  date: string;
+  raw: string;
+  ficheSpDescription: string;
+}
+export interface ContributionGenericNoCDTContent {
   messageBlockGenericNoCDT: string;
-  messageBlockGenericNoCDTUnextendedCC: string;
-};
+}
 
-type ContributionDocumentJsonFicheSp = ContributionDocumentJsonBasic & {
-  type: "fiche-sp";
-  ficheSpId: string;
-};
-
-type ContributionDocumentJsonCodeDuTravailReference =
-  ContributionDocumentJsonBasic & {
-    type: "cdt";
-    genericAnswerId: string;
-  };
-
-export type ContributionDocumentJson =
-  | ContributionDocumentJsonContent
-  | ContributionDocumentJsonGenericNoCDT
-  | ContributionDocumentJsonFicheSp
-  | ContributionDocumentJsonCodeDuTravailReference;
+export type ContributionContent =
+  | ContributionContentBase
+  | ContributionFicheSpContent
+  | ContributionGenericNoCDTContent;
 
 export interface ExportFullLinkedContent {
   breadcrumbs: Breadcrumb[];
@@ -83,7 +58,7 @@ type ContributionElasticDocumentBase = Omit<
   "breadcrumbs"
 > &
   ContributionMetadata &
-  ContributionDocumentJson &
+  ContributionContent &
   ExportContributionFullLinkedContent &
   ExportContributionInfo;
